refactor(ui): move service lifecycle actions out of model.go
Resolving what a key targets, expanding the dependency graph, running the
operation, and confirming the destructive ones is the largest theme in the
package and was spread across the whole file.

Declarations move unchanged.
